codeintel: Move commitgraph package into uploads
