+++ internal/codeintel/uploads/internal/store/types.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/keegancsmith/sqlf"
 
-	"github.com/sourcegraph/sourcegraph/internal/codeintel/commitgraph"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/uploads/internal/commitgraph"
 )
 
 type cteDefinition struct {
